massive additions to the `mod backtrace_hacks`.
In particular, I pulled in some `dladdr` code that allows me to map
from an address to the base address of the shared object that it is
associated with. This in turn allows me to make some sense out of the
instruction offsets that are embedded in the LLVM StackSize records.

(It turns out that those records on their own are actually not useful
in the end, so maybe I will throw away this code eventually. That or
try to figure out some way to extend what the LLVM backend is doing to
be a better match for my needs.)