stackmap status: querying ReturnAddress against StackMap yields sketchy results.
In particular, it seems like the call-site's computed according to the stack map
is the address of a move instruction that immediately *precedes* the actual call,
while, (as one would indeed expect), the ReturnAddress is whatever instruction
that *follows* the actual call.

I'm not so concerned about the latter, but I would like to better
understand what is happening with respect to the former.